requirements.txt
streamlit>=1.30.0
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
pyserial>=3.5
python-dotenv>=1.0.0
